pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.12",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.10",
    "langgraph-supervisor>=0.0.31",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
